requirements.txt
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
paho-mqtt>=1.6.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development tools
black>=23.7.0
flake8>=6.1.0
pylint>=2.17.5